Move method from resource to datasource file
